+++ elArca/elArca/Common/NoHighlightButtonStyle.swift
@@ -1,0 +1,16 @@
+//
+//  NoHighlightButtonStyle.swift
+//  elArca
+//
+//  Created by Fátima Figueroa on 30/11/25.
+//
+
+import SwiftUI
+
+struct NoHighlightButtonStyle: ButtonStyle {
+    func makeBody(configuration: Configuration) -> some View {
+        configuration.label
+            .opacity(configuration.isPressed ? 0.9 : 1)
+            .animation(.easeInOut(duration: 0.1), value: configuration.isPressed)
+    }
+}