+++ elArca/elArca/Data/Repository/ApiConfig.swift
@@ -7,6 +7,7 @@
 
 struct Api {
     static let base = "http://74.208.78.8:8080/"
+//  static let base = "http://localhost:8080/"
     struct routes {
         static let calendar = "calendar/"
         static let workshops = "workshop/"
@@ -14,6 +15,7 @@ struct Api {
         
         static let beneficiary = "beneficiary/"
         static let notifications = "notifications/"
+        static let attendance = "qr/validate/"
         //static let otherEndpoint = "otherEndpoint/"
         //Define the endpoints here as needed and then use them in the repositories.
     }