+++ elArca/elArca/Framework/Views/Workshop/WorkshopView.swift
@@ -20,7 +20,8 @@ struct WorkshopView: View {
                         .ignoresSafeArea()
                     
                     VStack(spacing: 0) {
-                        // Header with Title and Notification Bell (needs to update icons and add navbar)
+                        
+                        // Header con título y campana
                         HStack {
                             Texts(text: "Talleres", type: .header)
                                 .foregroundColor(.white)
@@ -33,9 +34,31 @@ struct WorkshopView: View {
                         .padding(.top, 20)
                         .padding(.bottom, 20)
                         
+                        // Search Bar
+                        HStack(spacing: 8) {
+                            TextInput(
+                                value: $viewModel.searchText,
+                                errorMessage: .constant(""),
+                                label: "",
+                                placeholder: "Buscar taller",
+                                type: .searchInput
+                            )
+                            
+//                            IconButtonAtom(imageName: "filter") {
+//                                print("Filtro Talleres")
+//                            }
+                        }
+                        .padding(.horizontal, 24)
+                        .padding(.bottom, 12)
+                        .onChange(of: viewModel.searchText) { _ in
+                            viewModel.filterWorkshops()
+                        }
+
+                        
                         // Contenido principal
                         mainContent
                     }
+
                 }
                 .edgesIgnoringSafeArea(.bottom)
             }